feat: DAH-2639 add no lottery required to mobile
