--- app/javascript/modules/listingDetailsAside/ListingDetailsAside.scss
@@ -1,21 +1,0 @@
-.fcfs-no-lottery {
-    --card-border-radius: rounded-none;
-    --card-background-color: var(--seeds-color-gray-200);
-    --card-footer-background-color: var(--seeds-color-gray-200);
-    --card-border-width: var(--seeds-border-0);
-    border-bottom: 1px solid var(--seeds-border-color);
-
-    .fcfs-no-lottery-header {
-        font-family: var(--seeds-font-alt-sans);
-        font-size: var(--seeds-type-heading-size-md);
-        font-weight: var(--seeds-font-weight-semibold);
-        line-height: var(--seeds-line-height-6);
-    }
-
-    .fcfs-no-lottery-section {
-        font-family: var(--seeds-font-sans);
-        font-size: var(--seeds-font-size-sm);
-        font-weight: var(--seeds-font-weight-regular);
-        line-height: var(--seeds-line-height-5);
-    }
-}
