+++ app/javascript/modules/listingDetailsAside/ListingDetailsAside.tsx
@@ -12,11 +12,9 @@ import { ListingDetailsWaitlist } from "./ListingDetailsWaitlist"
 import { ListingDetailsOpenHouses } from "./ListingDetailsOpenHouses"
 import { ListingDetailsSeeTheUnit } from "./ListingDetailsSeeTheUnit"
 import { useFeatureFlag } from "../../hooks/useFeatureFlag"
-import { getSfGovUrl, localizedFormat } from "../../util/languageUtil"
+import { localizedFormat } from "../../util/languageUtil"
 import { getHousingCounselorsPath } from "../../util/routeUtil"
-import { Card, Link } from "@bloom-housing/ui-seeds"
-import { CardFooter, CardHeader, CardSection } from "@bloom-housing/ui-seeds/src/blocks/Card"
-import "./ListingDetailsAside.scss"
+import { fcfsNoLotteryRequired } from "../noLotteryRequired/fcfsNoLotteryRequired"
 
 export interface ListingDetailsSidebarProps {
   listing: RailsListing
@@ -57,28 +55,6 @@ export const ListingDetailsAside = ({ listing, imageSrc }: ListingDetailsSidebar
       </LinkButton>
     </SidebarBlock>
   )
-
-  const fcfsNoLotteryRequired = () => {
-    const url = getSfGovUrl(
-      "https://www.sf.gov/step-by-step/buy-home-without-entering-lottery",
-      14246
-    )
-    return (
-      <Card className="fcfs-no-lottery">
-        <CardHeader className="font-bold fcfs-no-lottery-header">
-          {t("listings.fcfs.bmrSales.noLotteryRequired.header")}
-        </CardHeader>
-        <CardSection className="fcfs-no-lottery-section">
-          {t("listings.fcfs.bmrSales.noLotteryRequired.section")}
-        </CardSection>
-        <CardFooter className="ml-6 mb-6 underline fcfs-no-lottery-section">
-          <Link href={url} hideExternalLinkIcon={true}>
-            {t("listings.fcfs.bmrSales.noLotteryRequired.footer")}
-          </Link>
-        </CardFooter>
-      </Card>
-    )
-  }
 
   return (
     <ul>
